+++ docs/content/auto-imports.d.ts
@@ -7,4 +7,5 @@
 export {}
 declare global {
   const ElButton: typeof import('element-plus/es').ElButton
+  const ElMessageBox: typeof import('element-plus/es').ElMessageBox
 }