+++ docs/examples/components/basic/dialog/basic/index.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import type { EllDialogResult } from 'element-plus-lab'
+import type { EllOverlayResult } from 'element-plus-lab'
 import { ElMessage } from 'element-plus'
 import { useDialog } from 'element-plus-lab'
 
@@ -7,15 +7,15 @@ const { openDialog } = useDialog()
 
 async function openWithAsyncAwait() {
   try {
-    await openDialog({ content: '确认删除？' })
-    ElMessage.success('删除成功！')
+    await openDialog({ content: '确定要执行此操作吗？' })
+    ElMessage.success('已确认')
   }
   catch (error: unknown) {
-    // 明确 error 是 EllDialogResult 类型
-    const result = error as unknown as EllDialogResult
+    // 明确 error 是 EllOverlayResult 类型
+    const result = error as unknown as EllOverlayResult
     console.error(result)
     if (result.reason === 'cancel') {
-      ElMessage.info('取消删除')
+      ElMessage.info('已取消')
     }
     else if (result.reason === 'close') {
       ElMessage.info('关闭对话框')
@@ -27,14 +27,14 @@ async function openWithAsyncAwait() {
 }
 
 function openWithPromise() {
-  openDialog({ content: '确认删除？' })
+  openDialog({ content: '确定要执行此操作吗？' })
     .then(() => {
-      ElMessage.success('删除成功！')
+      ElMessage.success('已确认')
     })
     .catch((error) => {
       // 这里参考上面的 async/await 方式来处理错误，不再赘述
       console.error(error)
-      ElMessage.info('取消删除')
+      ElMessage.info('已取消')
     })
 }
 </script>