+++ packages/element-plus-lab/index.ts
@@ -1,2 +1,9 @@
 export { useDialog } from '../components/dialog'
+export { useDrawer } from '../components/drawer'
+
 export type * from '../components/dialog'
+export type * from '../components/drawer'
+export type {
+  EllOverlayCloseReason,
+  EllOverlayResult,
+} from '@element-plus-lab/utils'